[BTAT-2131] split integration test constants into separate files
