+++ it/controllers/CalculationControllerISpec.scala
@@ -16,7 +16,13 @@
 package controllers
 
 import enums.{Crystallised, Estimate}
-import helpers.IntegrationTestConstants._
+import assets.BaseIntegrationTestConstants._
+import assets.BusinessDetailsIntegrationTestConstants._
+import assets.CalcDataIntegrationTestConstants._
+import assets.FinancialTransactionsIntegrationTestConstants._
+import assets.PropertyDetailsIntegrationTestConstants._
+import assets.ReportDeadlinesIntegrationTestConstants._
+import assets.StatementsIntegrationTestConstants._
 import helpers.servicemocks._
 import helpers.{ComponentSpecBase, GenericStubMethods}
 import models.{CalculationDataErrorModel, CalculationDataModel, LastTaxCalculation, FinancialTransactionsModel}
@@ -51,16 +57,16 @@ class CalculationControllerISpec extends ComponentSpecBase with GenericStubMetho
 
         And("I wiremock stub a successful Get Last Estimated Tax Liability response")
         val lastTaxCalcResponse =
-          LastTaxCalculation(testCalcId, "2017-07-06T12:34:56.789Z", GetCalculationData.calculationDataSuccessWithEoYModel.totalIncomeTaxNicYtd, Estimate)
+          LastTaxCalculation(testCalcId, "2017-07-06T12:34:56.789Z", calculationDataSuccessWithEoYModel.totalIncomeTaxNicYtd, Estimate)
         IncomeTaxViewChangeStub.stubGetLastTaxCalc(testNino, testYear, lastTaxCalcResponse)
 
         And("I wiremock stub a successful Get CalculationData response")
-        val calcBreakdownResponse = GetCalculationData.calculationDataSuccessWithEoYModel
-        IncomeTaxViewChangeStub.stubGetCalcData(testNino, testCalcId, GetCalculationData.calculationDataSuccessWithEoyJson.toString())
+        val calcBreakdownResponse = calculationDataSuccessWithEoYModel
+        IncomeTaxViewChangeStub.stubGetCalcData(testNino, testCalcId, calculationDataSuccessWithEoyJson.toString())
 
-        getBizDeets(GetBusinessDetails.successResponse(testSelfEmploymentId))
+        getBizDeets(businessSuccessResponse(testSelfEmploymentId))
 
-        getPropDeets(GetPropertyDetails.successResponse())
+        getPropDeets(propertySuccessResponse())
 
         When(s"I call GET /report-quarterly/income-and-expenses/view/calculation/$testYear")
         val res = IncomeTaxViewChangeFrontend.getFinancialData(testYear)
@@ -119,21 +125,21 @@ class CalculationControllerISpec extends ComponentSpecBase with GenericStubMetho
 
           isAuthorisedUser(true)
           stubUserDetails()
-          getBizDeets(GetBusinessDetails.successResponse(testSelfEmploymentId))
-          getPropDeets(GetPropertyDetails.successResponse())
+          getBizDeets(businessSuccessResponse(testSelfEmploymentId))
+          getPropDeets(propertySuccessResponse())
 
           And("I wiremock stub a successful Get Last Estimated Tax Liability response")
           val lastTaxCalcResponse =
-            LastTaxCalculation(testCalcId, "2017-07-06T12:34:56.789Z", GetCalculationData.calculationDataSuccessWithEoYModel.totalIncomeTaxNicYtd, Crystallised)
+            LastTaxCalculation(testCalcId, "2017-07-06T12:34:56.789Z", calculationDataSuccessWithEoYModel.totalIncomeTaxNicYtd, Crystallised)
           IncomeTaxViewChangeStub.stubGetLastTaxCalc(testNino, testYear, lastTaxCalcResponse)
 
           And("I wiremock stub a successful Get CalculationData response")
-          val calcBreakdownResponse = GetCalculationData.calculationDataSuccessWithEoYModel
-          IncomeTaxViewChangeStub.stubGetCalcData(testNino, testCalcId, GetCalculationData.calculationDataSuccessWithEoyJson.toString())
+          val calcBreakdownResponse = calculationDataSuccessWithEoYModel
+          IncomeTaxViewChangeStub.stubGetCalcData(testNino, testCalcId, calculationDataSuccessWithEoyJson.toString())
 
           And("I wiremock stub a successful Unpaid Financial Transactions response")
-          val financialTransactions = GetFinancialTransactions.financialTransactionsJson(1000.0).as[FinancialTransactionsModel]
-          FinancialTransactionsStub.stubGetFinancialTransactions(testMtditid)(Status.OK, GetFinancialTransactions.financialTransactionsJson(1000.0))
+          val financialTransactions = financialTransactionsJson(1000.0).as[FinancialTransactionsModel]
+          FinancialTransactionsStub.stubGetFinancialTransactions(testMtditid)(Status.OK, financialTransactionsJson(1000.0))
 
           When(s"I call GET /report-quarterly/income-and-expenses/view/calculation/$testYear")
           val res = IncomeTaxViewChangeFrontend.getFinancialData(testYear)
@@ -194,21 +200,21 @@ class CalculationControllerISpec extends ComponentSpecBase with GenericStubMetho
 
           isAuthorisedUser(true)
           stubUserDetails()
-          getBizDeets(GetBusinessDetails.successResponse(testSelfEmploymentId))
-          getPropDeets(GetPropertyDetails.successResponse())
+          getBizDeets(businessSuccessResponse(testSelfEmploymentId))
+          getPropDeets(propertySuccessResponse())
 
           And("I wiremock stub a successful Get Last Estimated Tax Liability response")
           val lastTaxCalcResponse =
-            LastTaxCalculation(testCalcId, "2017-07-06T12:34:56.789Z", GetCalculationData.calculationDataSuccessWithEoYModel.totalIncomeTaxNicYtd, Crystallised)
+            LastTaxCalculation(testCalcId, "2017-07-06T12:34:56.789Z", calculationDataSuccessWithEoYModel.totalIncomeTaxNicYtd, Crystallised)
           IncomeTaxViewChangeStub.stubGetLastTaxCalc(testNino, testYear, lastTaxCalcResponse)
 
           And("I wiremock stub a successful Get CalculationData response")
-          val calcBreakdownResponse = GetCalculationData.calculationDataSuccessWithEoYModel
-          IncomeTaxViewChangeStub.stubGetCalcData(testNino, testCalcId, GetCalculationData.calculationDataSuccessWithEoyJson.toString())
+          val calcBreakdownResponse = calculationDataSuccessWithEoYModel
+          IncomeTaxViewChangeStub.stubGetCalcData(testNino, testCalcId, calculationDataSuccessWithEoyJson.toString())
 
           And("I wiremock stub a successful paid Financial Transactions response")
-          val financialTransactions = GetFinancialTransactions.financialTransactionsJson(0.0).as[FinancialTransactionsModel]
-          FinancialTransactionsStub.stubGetFinancialTransactions(testMtditid)(Status.OK, GetFinancialTransactions.financialTransactionsJson(0.0))
+          val financialTransactions = financialTransactionsJson(0.0).as[FinancialTransactionsModel]
+          FinancialTransactionsStub.stubGetFinancialTransactions(testMtditid)(Status.OK, financialTransactionsJson(0.0))
 
           When(s"I call GET /report-quarterly/income-and-expenses/view/calculation/$testYear")
           val res = IncomeTaxViewChangeFrontend.getFinancialData(testYear)
@@ -267,20 +273,20 @@ class CalculationControllerISpec extends ComponentSpecBase with GenericStubMetho
 
           isAuthorisedUser(true)
           stubUserDetails()
-          getBizDeets(GetBusinessDetails.successResponse(testSelfEmploymentId))
-          getPropDeets(GetPropertyDetails.successResponse())
+          getBizDeets(businessSuccessResponse(testSelfEmploymentId))
+          getPropDeets(propertySuccessResponse())
 
           And("I wiremock stub a successful Get Last Estimated Tax Liability response")
           val lastTaxCalcResponse =
-            LastTaxCalculation(testCalcId, "2017-07-06T12:34:56.789Z", GetCalculationData.calculationDataSuccessWithEoYModel.totalIncomeTaxNicYtd, Crystallised)
+            LastTaxCalculation(testCalcId, "2017-07-06T12:34:56.789Z", calculationDataSuccessWithEoYModel.totalIncomeTaxNicYtd, Crystallised)
           IncomeTaxViewChangeStub.stubGetLastTaxCalc(testNino, testYear, lastTaxCalcResponse)
 
           And("I wiremock stub a successful Get CalculationData response")
-          val calcBreakdownResponse = GetCalculationData.calculationDataSuccessWithEoYModel
-          IncomeTaxViewChangeStub.stubGetCalcData(testNino, testCalcId, GetCalculationData.calculationDataSuccessWithEoyJson.toString())
+          val calcBreakdownResponse = calculationDataSuccessWithEoYModel
+          IncomeTaxViewChangeStub.stubGetCalcData(testNino, testCalcId, calculationDataSuccessWithEoyJson.toString())
 
           And("I wiremock stub an errored Financial Transactions response")
-          FinancialTransactionsStub.stubGetFinancialTransactions(testMtditid)(Status.OK, GetFinancialTransactions.financialTransactionsSingleErrorJson)
+          FinancialTransactionsStub.stubGetFinancialTransactions(testMtditid)(Status.OK, financialTransactionsSingleErrorJson)
 
           When(s"I call GET /report-quarterly/income-and-expenses/view/calculation/$testYear")
           val res = IncomeTaxViewChangeFrontend.getFinancialData(testYear)
@@ -310,16 +316,16 @@ class CalculationControllerISpec extends ComponentSpecBase with GenericStubMetho
 
         And("I wiremock stub a successful Get Last Estimated Tax Liability response")
         val lastTaxCalcResponse =
-          LastTaxCalculation(testCalcId, "2017-07-06T12:34:56.789Z", GetCalculationData.calculationDataSuccessModel.totalIncomeTaxNicYtd, Estimate)
+          LastTaxCalculation(testCalcId, "2017-07-06T12:34:56.789Z", calculationDataSuccessModel.totalIncomeTaxNicYtd, Estimate)
         IncomeTaxViewChangeStub.stubGetLastTaxCalc(testNino, testYear, lastTaxCalcResponse)
 
         And("I wiremock stub a successful Get CalculationData response")
-        val calcBreakdownResponse = GetCalculationData.calculationDataSuccessModel
-        IncomeTaxViewChangeStub.stubGetCalcData(testNino, testCalcId, GetCalculationData.calculationDataSuccessJson.toString())
+        val calcBreakdownResponse = calculationDataSuccessModel
+        IncomeTaxViewChangeStub.stubGetCalcData(testNino, testCalcId, calculationDataSuccessJson.toString())
 
-        getBizDeets(GetBusinessDetails.successResponse(testSelfEmploymentId))
+        getBizDeets(businessSuccessResponse(testSelfEmploymentId))
 
-        getPropDeets(GetPropertyDetails.successResponse())
+        getPropDeets(propertySuccessResponse())
 
         When(s"I call GET /report-quarterly/income-and-expenses/view/calculation/$testYear")
         val res = IncomeTaxViewChangeFrontend.getFinancialData(testYear)
@@ -383,18 +389,18 @@ class CalculationControllerISpec extends ComponentSpecBase with GenericStubMetho
 
         And("a successful Get Last Estimated Tax Liability response via wiremock stub")
         val lastTaxCalcResponse =
-          LastTaxCalculation(testCalcId, "2017-07-06T12:34:56.789Z", GetCalculationData.calculationDataSuccessWithEoYModel.totalIncomeTaxNicYtd, Estimate)
+          LastTaxCalculation(testCalcId, "2017-07-06T12:34:56.789Z", calculationDataSuccessWithEoYModel.totalIncomeTaxNicYtd, Estimate)
         IncomeTaxViewChangeStub.stubGetLastTaxCalc(testNino, testYear, lastTaxCalcResponse)
 
-        And("I wiremock stub an erroneous GetCalculationData response")
-        val calc = GetCalculationData.calculationDataErrorModel
+        And("I wiremock stub an erroneous response")
+        val calc = calculationDataErrorModel
         val calculationResponse = CalculationDataErrorModel(calc.code, calc.message)
 
         IncomeTaxViewChangeStub.stubGetCalcError(testNino, testCalcId, calculationResponse)
 
-        getBizDeets(GetBusinessDetails.successResponse(testSelfEmploymentId))
+        getBizDeets(businessSuccessResponse(testSelfEmploymentId))
 
-        getPropDeets(GetPropertyDetails.successResponse())
+        getPropDeets(propertySuccessResponse())
 
         When(s"I make a call to GET /report-quarterly/income-and-expenses/view/calculation/$testYear")
         val res = IncomeTaxViewChangeFrontend.getFinancialData(testYear)
@@ -410,7 +416,7 @@ class CalculationControllerISpec extends ComponentSpecBase with GenericStubMetho
         res should have(
           httpStatus(OK),
           pageTitle("Tax year: 2017 to 2018"),
-          elementTextByID("inYearEstimateHeading")(s"Current estimate: ${GetCalculationData.calculationDataSuccessWithEoYModel.totalIncomeTaxNicYtd.toCurrencyString}"),
+          elementTextByID("inYearEstimateHeading")(s"Current estimate: ${calculationDataSuccessWithEoYModel.totalIncomeTaxNicYtd.toCurrencyString}"),
           elementTextByID("inYearP1")("This is an estimate of the tax you owe from 6 April 2017 to 6 July 2017."),
           isElementVisibleById("inYearCalcBreakdown")(expectedValue = false)
         )
@@ -430,9 +436,9 @@ class CalculationControllerISpec extends ComponentSpecBase with GenericStubMetho
         And("a No Data Found response from Get Last Estimated Tax Liability via wiremock stub")
         IncomeTaxViewChangeStub.stubGetLastCalcNoData(testNino, testYear)
 
-        getBizDeets(GetBusinessDetails.successResponse(testSelfEmploymentId))
+        getBizDeets(businessSuccessResponse(testSelfEmploymentId))
 
-        getPropDeets(GetPropertyDetails.successResponse())
+        getPropDeets(propertySuccessResponse())
 
         When(s"I make a call to GET /report-quarterly/income-and-expenses/view/calculation/$testYear ")
         val res = IncomeTaxViewChangeFrontend.getFinancialData(testYear)
@@ -465,9 +471,9 @@ class CalculationControllerISpec extends ComponentSpecBase with GenericStubMetho
         And("an Error Response response from Get Last Estimated Tax Liability via wiremock stub")
         IncomeTaxViewChangeStub.stubGetLastCalcError(testNino, testYear)
 
-        getBizDeets(GetBusinessDetails.successResponse(testSelfEmploymentId))
+        getBizDeets(businessSuccessResponse(testSelfEmploymentId))
 
-        getPropDeets(GetPropertyDetails.successResponse())
+        getPropDeets(propertySuccessResponse())
 
         When(s"I make a call to GET /report-quarterly/income-and-expenses/view/calculation/$testYear ")
         val res = IncomeTaxViewChangeFrontend.getFinancialData(testYear)
