+++ it/controllers/EstimatesControllerISpec.scala
@@ -15,9 +15,12 @@
  */
 package controllers
 
+import assets.BaseIntegrationTestConstants._
+import assets.BusinessDetailsIntegrationTestConstants._
+import assets.CalcDataIntegrationTestConstants._
+import assets.PropertyDetailsIntegrationTestConstants._
 import config.FrontendAppConfig
 import enums.{Crystallised, Estimate}
-import helpers.IntegrationTestConstants._
 import helpers.servicemocks._
 import helpers.{ComponentSpecBase, GenericStubMethods}
 import models.LastTaxCalculation
@@ -37,8 +40,8 @@ class EstimatesControllerISpec extends ComponentSpecBase with GenericStubMethods
 
         isAuthorisedUser(true)
         stubUserDetails()
-        getBizDeets(GetBusinessDetails.successResponse(testSelfEmploymentId))
-        getPropDeets(GetPropertyDetails.successResponse())
+        getBizDeets(businessSuccessResponse(testSelfEmploymentId))
+        getPropDeets(propertySuccessResponse())
 
         When(s"I call GET /report-quarterly/income-and-expenses/view/estimates")
         val res = IncomeTaxViewChangeFrontend.getEstimates
@@ -59,17 +62,17 @@ class EstimatesControllerISpec extends ComponentSpecBase with GenericStubMethods
           appConfig.features.estimatesEnabled(true)
           isAuthorisedUser(true)
           stubUserDetails()
-          getBizDeets(GetBusinessDetails.successResponse(testSelfEmploymentId))
-          getPropDeets(GetPropertyDetails.successResponse())
+          getBizDeets(businessSuccessResponse(testSelfEmploymentId))
+          getPropDeets(propertySuccessResponse())
 
           And("I wiremock stub a successful Get Last Estimated Tax Liability response")
           val lastTaxCalcResponse =
-            LastTaxCalculation(testCalcId, "2017-07-06T12:34:56.789Z", GetCalculationData.calculationDataSuccessWithEoYModel.totalIncomeTaxNicYtd, Estimate)
+            LastTaxCalculation(testCalcId, "2017-07-06T12:34:56.789Z", calculationDataSuccessWithEoYModel.totalIncomeTaxNicYtd, Estimate)
           IncomeTaxViewChangeStub.stubGetLastTaxCalc(testNino, testYear, lastTaxCalcResponse)
 
           And("I wiremock stub a successful Get CalculationData response")
-          val calcBreakdownResponse = GetCalculationData.calculationDataSuccessWithEoYModel
-          IncomeTaxViewChangeStub.stubGetCalcData(testNino, testCalcId, GetCalculationData.calculationDataSuccessWithEoyJson.toString())
+          val calcBreakdownResponse = calculationDataSuccessWithEoYModel
+          IncomeTaxViewChangeStub.stubGetCalcData(testNino, testCalcId, calculationDataSuccessWithEoyJson.toString())
 
           When(s"I call GET /report-quarterly/income-and-expenses/view/estimates")
           val res = IncomeTaxViewChangeFrontend.getEstimates
@@ -94,21 +97,21 @@ class EstimatesControllerISpec extends ComponentSpecBase with GenericStubMethods
           appConfig.features.estimatesEnabled(true)
           isAuthorisedUser(true)
           stubUserDetails()
-          getBizDeets(GetBusinessDetails.multipleSuccessResponse(testSelfEmploymentId, otherTestSelfEmploymentId))
-          getPropDeets(GetPropertyDetails.successResponse())
+          getBizDeets(multipleSuccessResponse(testSelfEmploymentId, otherTestSelfEmploymentId))
+          getPropDeets(propertySuccessResponse())
 
           And("I wiremock stub a successful Get Last Estimated Tax Liability response")
           val lastTaxCalcResponse =
-            LastTaxCalculation(testCalcId, "2017-07-06T12:34:56.789Z", GetCalculationData.calculationDataSuccessWithEoYModel.totalIncomeTaxNicYtd, Estimate)
+            LastTaxCalculation(testCalcId, "2017-07-06T12:34:56.789Z", calculationDataSuccessWithEoYModel.totalIncomeTaxNicYtd, Estimate)
           val lastTaxCalcResponse2 =
-            LastTaxCalculation(testCalcId2, "2017-07-06T12:34:56.789Z", GetCalculationData.calculationDataSuccessModel.totalIncomeTaxNicYtd, Estimate)
+            LastTaxCalculation(testCalcId2, "2017-07-06T12:34:56.789Z", calculationDataSuccessModel.totalIncomeTaxNicYtd, Estimate)
           IncomeTaxViewChangeStub.stubGetLastTaxCalc(testNino, testYear, lastTaxCalcResponse)
           IncomeTaxViewChangeStub.stubGetLastTaxCalc(testNino, testYearPlusOne, lastTaxCalcResponse2)
 
           And("I wiremock stub a successful Get CalculationData response")
-          val calcBreakdownResponse = GetCalculationData.calculationDataSuccessWithEoYModel
-          IncomeTaxViewChangeStub.stubGetCalcData(testNino, testCalcId, GetCalculationData.calculationDataSuccessWithEoyJson.toString())
-          IncomeTaxViewChangeStub.stubGetCalcData(testNino, testCalcId2, GetCalculationData.calculationDataSuccessWithEoyJson.toString())
+          val calcBreakdownResponse = calculationDataSuccessWithEoYModel
+          IncomeTaxViewChangeStub.stubGetCalcData(testNino, testCalcId, calculationDataSuccessWithEoyJson.toString())
+          IncomeTaxViewChangeStub.stubGetCalcData(testNino, testCalcId2, calculationDataSuccessWithEoyJson.toString())
 
           When(s"I call GET /report-quarterly/income-and-expenses/view/estimates")
           val res = IncomeTaxViewChangeFrontend.getEstimates
@@ -138,31 +141,31 @@ class EstimatesControllerISpec extends ComponentSpecBase with GenericStubMethods
           appConfig.features.estimatesEnabled(true)
           isAuthorisedUser(true)
           stubUserDetails()
-          getBizDeets(GetBusinessDetails.multipleSuccessResponse(testSelfEmploymentId, otherTestSelfEmploymentId))
-          getPropDeets(GetPropertyDetails.successResponse())
+          getBizDeets(multipleSuccessResponse(testSelfEmploymentId, otherTestSelfEmploymentId))
+          getPropDeets(propertySuccessResponse())
 
           And("I wiremock stub a successful Get Last Estimated Tax Liability response")
           val crystallisedLastTaxCalcResponse =
             LastTaxCalculation(
               testCalcId,
               "2017-07-06T12:34:56.789Z",
-              GetCalculationData.calculationDataSuccessWithEoYModel.totalIncomeTaxNicYtd,
+              calculationDataSuccessWithEoYModel.totalIncomeTaxNicYtd,
               Crystallised
             )
           val lastTaxCalcResponse =
             LastTaxCalculation(
               testCalcId2,
               "2017-07-06T12:34:56.789Z",
-              GetCalculationData.calculationDataSuccessModel.totalIncomeTaxNicYtd,
+              calculationDataSuccessModel.totalIncomeTaxNicYtd,
               Estimate
             )
           IncomeTaxViewChangeStub.stubGetLastTaxCalc(testNino, testYear, crystallisedLastTaxCalcResponse)
           IncomeTaxViewChangeStub.stubGetLastTaxCalc(testNino, testYearPlusOne, lastTaxCalcResponse)
 
           And("I wiremock stub a successful Get CalculationData response")
-          val calcBreakdownResponse = GetCalculationData.calculationDataSuccessWithEoYModel
-          IncomeTaxViewChangeStub.stubGetCalcData(testNino, testCalcId, GetCalculationData.calculationDataSuccessWithEoyJson.toString())
-          IncomeTaxViewChangeStub.stubGetCalcData(testNino, testCalcId2, GetCalculationData.calculationDataSuccessWithEoyJson.toString())
+          val calcBreakdownResponse = calculationDataSuccessWithEoYModel
+          IncomeTaxViewChangeStub.stubGetCalcData(testNino, testCalcId, calculationDataSuccessWithEoyJson.toString())
+          IncomeTaxViewChangeStub.stubGetCalcData(testNino, testCalcId2, calculationDataSuccessWithEoyJson.toString())
 
 
           When(s"I call GET /report-quarterly/income-and-expenses/view/estimates")
@@ -189,22 +192,22 @@ class EstimatesControllerISpec extends ComponentSpecBase with GenericStubMethods
           appConfig.features.estimatesEnabled(true)
           isAuthorisedUser(true)
           stubUserDetails()
-          getBizDeets(GetBusinessDetails.successResponse(testSelfEmploymentId))
-          getPropDeets(GetPropertyDetails.successResponse())
+          getBizDeets(businessSuccessResponse(testSelfEmploymentId))
+          getPropDeets(propertySuccessResponse())
 
           And("I wiremock stub a successful Get Last Estimated Tax Liability response")
           val lastTaxCalcResponse =
             LastTaxCalculation(
               testCalcId,
               "2017-07-06T12:34:56.789Z",
-              GetCalculationData.calculationDataSuccessWithEoYModel.totalIncomeTaxNicYtd,
+              calculationDataSuccessWithEoYModel.totalIncomeTaxNicYtd,
               Crystallised
             )
           IncomeTaxViewChangeStub.stubGetLastTaxCalc(testNino, testYear, lastTaxCalcResponse)
 
           And("I wiremock stub a successful Get CalculationData response")
-          val calcBreakdownResponse = GetCalculationData.calculationDataSuccessWithEoYModel
-          IncomeTaxViewChangeStub.stubGetCalcData(testNino, testCalcId, GetCalculationData.calculationDataSuccessWithEoyJson.toString())
+          val calcBreakdownResponse = calculationDataSuccessWithEoYModel
+          IncomeTaxViewChangeStub.stubGetCalcData(testNino, testCalcId, calculationDataSuccessWithEoyJson.toString())
 
 
           When(s"I call GET /report-quarterly/income-and-expenses/view/estimates")
