+++ it/controllers/StatementsControllerISpec.scala
@@ -16,8 +16,9 @@
 
 package controllers
 
+import assets.BaseIntegrationTestConstants._
+import assets.StatementsIntegrationTestConstants._
 import config.FrontendAppConfig
-import helpers.IntegrationTestConstants._
 import helpers.servicemocks._
 import helpers.{ComponentSpecBase, GenericStubMethods}
 import play.api.http.Status
@@ -41,7 +42,7 @@ class StatementsControllerISpec extends ComponentSpecBase with ImplicitDateForma
         isAuthorisedUser(true)
 
         And("I wiremock stub a successful Get Financial Transactions response")
-        val statementResponse = Json.toJson(GetStatementsData.singleFinancialTransactionsModel)
+        val statementResponse = Json.toJson(singleFinancialTransactionsModel)
         FinancialTransactionsStub.stubGetFinancialTransactions(testMtditid)(Status.OK, statementResponse)
 
         When(s"I call GET /report-quarterly/income-and-expenses/view/statements")
@@ -67,7 +68,7 @@ class StatementsControllerISpec extends ComponentSpecBase with ImplicitDateForma
             isAuthorisedUser(true)
 
             And("I wiremock stub a successful Get Financial Transactions response")
-            val statementResponse = Json.toJson(GetStatementsData.singleFinancialTransactionsModel)
+            val statementResponse = Json.toJson(singleFinancialTransactionsModel)
             FinancialTransactionsStub.stubGetFinancialTransactions(testMtditid)(Status.OK, statementResponse)
 
             When(s"I call GET /report-quarterly/income-and-expenses/view/statements")
@@ -77,13 +78,13 @@ class StatementsControllerISpec extends ComponentSpecBase with ImplicitDateForma
             FinancialTransactionsStub.verifyGetFinancialTransactions(testMtditid)
 
             Then("The view should have the correct headings and single statement")
-            val model = GetStatementsData.singleChargeTransactionModel
+            val model = singleChargeTransactionModel
             res should have(
               httpStatus(OK),
               pageTitle("Income Tax Statement"),
               elementTextByID(s"$testYear-tax-year")(s"Tax year: ${testYearInt - 1}-$testYear"),
               elementTextByID(s"$testYear-still-to-pay")(s"Still to pay: ${model.outstandingAmount.get.toCurrencyString}"),
-              elementTextByID(s"$testYear-charge")(GetStatementsData.charge2018.amount.get.toCurrencyString),
+              elementTextByID(s"$testYear-charge")(charge2018.amount.get.toCurrencyString),
               isElementVisibleById("earlier-statements")(true),
               isElementVisibleById(s"$testYear-paid-0")(false)
             )
@@ -100,7 +101,7 @@ class StatementsControllerISpec extends ComponentSpecBase with ImplicitDateForma
             isAuthorisedUser(true)
 
             And("I wiremock stub a successful Get Financial Transactions response")
-            val statementResponse = Json.toJson(GetStatementsData.singleFTModel1charge2payments)
+            val statementResponse = Json.toJson(singleFTModel1charge2payments)
             FinancialTransactionsStub.stubGetFinancialTransactions(testMtditid)(Status.OK, statementResponse)
 
             When(s"I call GET /report-quarterly/income-and-expenses/view/statements")
@@ -110,20 +111,20 @@ class StatementsControllerISpec extends ComponentSpecBase with ImplicitDateForma
             FinancialTransactionsStub.verifyGetFinancialTransactions(testMtditid)
 
             Then("The view should have the correct headings and single statement")
-            val statement1Model = GetStatementsData.singleChargeTransactionModel
-            val statement2Model = GetStatementsData.singleCharge2PaymentsTransactionModel
-            val payment = GetStatementsData.payment2019
-            val payment2 = GetStatementsData.otherPayment2019
+            val statement1Model = singleChargeTransactionModel
+            val statement2Model = singleCharge2PaymentsTransactionModel
+            val payment = payment2019
+            val payment2 = otherPayment2019
             res should have(
               httpStatus(OK),
               pageTitle("Income Tax Statement"),
               elementTextByID(s"$testYear-tax-year")(s"Tax year: ${testYearInt - 1}-$testYear"),
               elementTextByID(s"$testYear-still-to-pay")(s"Still to pay: ${statement1Model.outstandingAmount.get.toCurrencyString}"),
-              elementTextByID(s"$testYear-charge")(GetStatementsData.charge2018.amount.get.toCurrencyString),
+              elementTextByID(s"$testYear-charge")(charge2018.amount.get.toCurrencyString),
               isElementVisibleById(s"$testYear-paid-0")(false),
               elementTextByID(s"$testYearPlusOne-tax-year")(s"Tax year: ${testYearPlusOneInt - 1}-$testYearPlusOne"),
               elementTextByID(s"$testYearPlusOne-still-to-pay")(s"Still to pay: ${statement2Model.outstandingAmount.get.toCurrencyString}"),
-              elementTextByID(s"$testYearPlusOne-charge")(GetStatementsData.charge2019.amount.get.toCurrencyString),
+              elementTextByID(s"$testYearPlusOne-charge")(charge2019.amount.get.toCurrencyString),
               elementTextByID(s"$testYearPlusOne-paid-0")(s"You paid " + payment.paymentAmount.get.toCurrencyString + " on " + payment.clearingDate.get.toShortDate),
               elementTextByID(s"$testYearPlusOne-paid-1")(s"You paid " + payment2.paymentAmount.get.toCurrencyString + " on " + payment2.clearingDate.get.toShortDate),
               isElementVisibleById("earlier-statements")(true)
@@ -141,7 +142,7 @@ class StatementsControllerISpec extends ComponentSpecBase with ImplicitDateForma
             isAuthorisedUser(true)
 
             And("I wiremock stub a successful Get Financial Transactions response")
-            val statementResponse = Json.toJson(GetStatementsData.emptyFTModel)
+            val statementResponse = Json.toJson(emptyFTModel)
             FinancialTransactionsStub.stubGetFinancialTransactions(testMtditid)(Status.OK, statementResponse)
 
             When(s"I call GET /report-quarterly/income-and-expenses/view/statements")
